Explode inputs module, Checksum.as_aria and File.checksum
- Exploded inputs module into many submodules
- Checksum gets an `as_aria` str prop
- File gets a `checksum`
