+++ src/offspot_config/inputs/ways.py
@@ -1,0 +1,1 @@
+WAYS = ("direct", "bztar", "gztar", "tar", "xztar", "zip")
